+++ src/rules/function-whitelist/index.js
@@ -17,7 +17,7 @@ export const messages = ruleMessages(ruleName, {
   rejected: (name) => `Unexpected function "${name}"`,
 })
 
-export default function (whitelistInput) {
+function rule(whitelistInput) {
   const whitelist = [].concat(whitelistInput)
   return (root, result) => {
     const validOptions = validateOptions(result, ruleName, {
@@ -42,3 +42,7 @@ export default function (whitelistInput) {
     })
   }
 }
+
+rule.primaryOptionArray = true
+
+export default rule